fix(commands): Refactor "purge" to fix a bug
Changed "purge" to no longer rely on unwraps with default values. Rust is the coolest programming language ever: the fact I can use returns to return from the function in a match statement that's assigning a variable, AND THEN use no return statement to specify what I want the variable to be assigned to?? I'm in shock at how beautiful this is.